Divide get repository list
This will divide to get repository list.
